Cleaned up code, properly renamed stack to dynamicarray
